Remove storage of matching results
Storing which tracks and cells were matched together depended on
the order of the tracks as passed in. This was potentially misleading
and could introduce errors if we try to evaluate on different ground
truth with the same parameters - it would not be rerun. I can rewrite
in a safe way where we store the actual tracks if necessary, but
the matching isn't very expensive if we just do it once per
configuration.