Remove csv_fields and read_dims from TrackSource
Instead of relying on multiple different possible ways for the user
to pass in the format of the tracks file, use the header of the tracks
file to determine order of values.
If no header is present, the node assumes an order of values. Both the
header values and the default assumed order are documented.